requirements.txt
black==25.1.0
click==8.1.8
colorama==0.4.6
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
